requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
arrow==1.3.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.3
binaryornot==0.4.4
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cookiecutter==2.6.0
crispy-bootstrap5==2025.4
cryptography==44.0.2
decouple==0.0.7
diff-match-patch==20241021
Django==4.2.20
django-allauth==65.7.0
django-appconf==1.1.0
django-browser-reload==1.18.0
django-crispy-forms==2.3
django-filter==25.1
django-geojson==4.2.0
django-humanize==0.1.2
django-import-export==4.3.7
django-leaflet==0.31.0
django-recaptcha==4.1.0
django-select2==8.2.3
django-simple-history==3.8.0
django-tailwind==3.6.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
dotenv==0.9.9
Faker==37.1.0
frozenlist==1.5.0
humanize==4.12.2
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.2
oauthlib==3.2.2
pillow==11.2.1
propcache==0.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
tqdm==4.67.1
twilio==9.5.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
whitenoise==6.9.0
yarl==1.19.0
zipp==3.21.0
gunicorn
psycopg2-binary

dotenv~=0.9.9
python-dotenv~=1.1.0
docx~=0.2.4
xhtml2pdf~=0.2.17
pandas~=2.2.3
celery~=5.5.1